+++ builder/builder_test.go
@@ -1,4 +1,4 @@
-package commands
+package builder
 
 import (
 	"testing"
